Ensure condition parser tests run in CI
