+++ dubbo-cluster/src/test/java/org/apache/dubbo/rpc/cluster/router/condition/config/ConditionStateRouterTestV31.java
@@ -116,39 +116,6 @@ public class ConditionStateRouterTestV31 {
     }
 
     @Test
-    public void testRejectNonMappingRule() {
-        for (String rawRule : Arrays.asList("", " ", "# comment", "---", "null", "[]", "condition")) {
-            IllegalArgumentException exception =
-                    Assertions.assertThrows(IllegalArgumentException.class, () -> ConditionRuleParser.parse(rawRule));
-            Assertions.assertEquals("Condition router rule must be a YAML mapping.", exception.getMessage());
-        }
-    }
-
-    @Test
-    public void testCommentOnlyRuleUpdatePreservesPreviousRule() {
-        String rawRule = "configVersion: v3.1\n" + "scope: service\n"
-                + "key: com.foo.BarService\n"
-                + "force: true\n"
-                + "runtime: true\n"
-                + "enabled: true\n"
-                + "conditions:\n"
-                + "  - from:\n"
-                + "      match: env=gray\n"
-                + "    to:\n"
-                + "      - match: env!=gray\n"
-                + "        weight: 100";
-
-        ServiceStateRouter<String> router = new ServiceStateRouter<>(
-                URL.valueOf("consumer://127.0.0.1/com.foo.BarService?env=gray&region=beijing"));
-        router.process(new ConfigChangedEvent("com.foo.BarService", "", rawRule, ConfigChangeType.ADDED));
-        Assertions.assertTrue(router.isForce());
-
-        router.process(new ConfigChangedEvent("com.foo.BarService", "", "# comment", ConfigChangeType.MODIFIED));
-
-        Assertions.assertTrue(router.isForce());
-    }
-
-    @Test
     public void testMultiplyConditionRoute() {
 
         String rawRule = "configVersion: v3.1\n" + "scope: service\n"
